Refactor UI components for shared partition status display
This PR extracts shared UI components for partition status display:

- Extract PartitionListSelector from AssetPartitionList for reuse
- Extract PartitionStatusCheckboxes from AssetPartitionStatusCheckboxes
- Refactor AssetPartitionList to use shared PartitionListSelector
- Refactor AssetPartitionStatusCheckboxes to use shared checkboxes
- Update client.json configuration

These components will be reused by asset check partition UI in the next PR.
